+++ tester/upnp_tester.c
@@ -48,7 +48,7 @@ static void upnp_check_ipaddress(void) {
 	addr = linphone_core_get_upnp_external_ipaddress(lc_upnp->lc);
 	BC_ASSERT_PTR_NOT_NULL(addr);
 	if (addr!=NULL) {
-		BC_ASSERT_GREATER(strlen(addr),7,int,"%d");
+		BC_ASSERT_GREATER((int)strlen(addr),7,int,"%d");
 	}
 	linphone_core_manager_destroy(lc_upnp);
 }
